+++ app/src/main/kotlin/org/fossify/messages/dialogs/InvalidNumberDialog.kt
@@ -12,7 +12,7 @@ class InvalidNumberDialog(val activity: BaseSimpleActivity, val text: String) {
         }
 
         activity.getAlertDialogBuilder()
-            .setPositiveButton(org.fossify.commons.R.string.ok) { _, _ -> { } }
+            .setPositiveButton(org.fossify.commons.R.string.ok) { _, _ -> }
             .apply {
                 activity.setupDialogStuff(binding.root, this)
             }