+++ app/src/main/kotlin/org/fossify/messages/helpers/MessagesImporter.kt
@@ -96,7 +96,8 @@ class MessagesImporter(private val activity: SimpleActivity) {
                         messagesFailed++
                     }
                 }
-                messageWriter.fixCoversationDates()
+
+                messageWriter.fixConversationDates()
                 refreshMessages()
             } catch (e: Exception) {
                 activity.showErrorToast(e)
@@ -154,7 +155,10 @@ class MessagesImporter(private val activity: SimpleActivity) {
                 refreshMessages()
             }
             when {
-                messagesFailed > 0 && messagesImported > 0 -> activity.toast(org.fossify.commons.R.string.importing_some_entries_failed)
+                messagesFailed > 0 && messagesImported > 0 -> {
+                    activity.toast(org.fossify.commons.R.string.importing_some_entries_failed)
+                }
+
                 messagesFailed > 0 -> activity.toast(org.fossify.commons.R.string.importing_failed)
                 else -> activity.toast(org.fossify.commons.R.string.importing_successful)
             }
@@ -211,12 +215,8 @@ class MessagesImporter(private val activity: SimpleActivity) {
 
     private fun isXmlMimeType(mimeType: String): Boolean {
         return mimeType.equals("application/xml", ignoreCase = true) || mimeType.equals(
-            "text/xml",
+            other = "text/xml",
             ignoreCase = true
         )
-    }
-
-    private fun isJsonMimeType(mimeType: String): Boolean {
-        return mimeType.equals("application/json", ignoreCase = true)
     }
 }