+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 Changelog
 ==========
 
+Version 1.1.3 *(2025-01-05)*
+----------------------------
+
+* Fixed issues with conversation date update (https://github.com/FossifyOrg/Messages/issues/225, https://github.com/FossifyOrg/Messages/issues/274)
+* Added more translations
+
+Version 1.1.2 *(2025-01-05)*
+----------------------------
+
+* Fixed issues with conversation date update (https://github.com/FossifyOrg/Messages/issues/225, https://github.com/FossifyOrg/Messages/issues/274)
+* Added more translations
+
+Version 1.1.1 *(2025-01-04)*
+----------------------------
+
+* Improved third party SMS/MMS intent parsing (https://github.com/FossifyOrg/Messages/issues/217, https://github.com/FossifyOrg/Messages/issues/243)
+* Modified short code check to exclude emails (https://github.com/FossifyOrg/Messages/issues/115)
+* Fixed issue with messages draft deletion (https://github.com/FossifyOrg/Messages/issues/13)
+* Fixed multiple toast errors for MMS messages (https://github.com/FossifyOrg/Messages/issues/70, https://github.com/FossifyOrg/Messages/issues/262)
+* Fixed some layout issues in message thread (https://github.com/FossifyOrg/Messages/issues/135)
+* Other minor bug fixes and improvements
+* Added more translations
+
+
 Version 1.1.0 *(2024-12-27)*
 ----------------------------